+++ client/src/Pages/Order/AddressForm/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import styled from "styled-components";
-import { AddressType, DestinationType, UserType } from "@/shared/type";
+import { AddressType, DestinationType } from "@/shared/type";
 import Button from "@/Components/Button";
 import useInput from "@/hooks/useInput";
 import useValidation from "@/hooks/useValidation";
@@ -10,34 +10,70 @@ import Address from "@/Components/Address";
 
 import { validatePhoneNumber, VALIDATION_ERR_MSG } from "@/utils/validations";
 import { gap } from "@/styles/theme";
+import { patchDestination, postDestination } from "@/api/destinations";
+import { convertToNumber, convertToPhoneNumber } from "@/utils/util";
 
-export type AddressFormProps = {
-  address?: DestinationType;
-  user?: UserType;
-};
+export interface AddressFormProps {
+  addressToEdit?: DestinationType;
+  gotoBack: Function;
+  refetch: Function;
+}
 
-const AddressForm = ({ address: addressToEdit, user }: AddressFormProps) => {
-  const name = useInput("");
+const AddressForm = ({
+  addressToEdit,
+  gotoBack,
+  refetch,
+}: AddressFormProps) => {
+  const addressee = useInput(addressToEdit?.addressee || "");
   const nameValidation = useValidation((name: string) => !!name.length);
-  const addressName = useInput("");
+  const addressName = useInput(addressToEdit?.name || "");
   const addressNameValidation = useValidation(
     (addressName: string) => !!addressName.length
   );
-  const phone = useInput("");
+  const phone = useInput(
+    addressToEdit?.phoneNumber || "",
+    convertToPhoneNumber
+  );
   const phoneValidation = useValidation(validatePhoneNumber);
 
   const [address, setAddress] = useState<AddressType>({
-    postCode: "",
-    address: "",
-    detailAddress: "",
+    address: addressToEdit?.address || "",
+    postCode: addressToEdit?.postCode || "",
+    detailAddress: addressToEdit?.detailAddress || "",
   });
 
   const handleChangeAddress = (address: DestinationType) => {
     setAddress(address);
   };
 
-  const handleSubmit = () => {
-    //TODO: API 요청 보낸 후 리디렉션
+  const handleSubmit = async () => {
+    try {
+      if (!addressToEdit) {
+        await postDestination({
+          data: {
+            ...address,
+            name: addressName.value,
+            addressee: addressee.value,
+            phoneNumber: convertToNumber(phone.value),
+          },
+        });
+      } else {
+        await patchDestination({
+          id: addressToEdit.id,
+          data: {
+            ...address,
+            name: addressName.value,
+            addressee: addressee.value,
+            phoneNumber: convertToNumber(phone.value),
+            isDefault: addressToEdit.isDefault,
+          },
+        });
+      }
+    } catch (error) {
+    } finally {
+      gotoBack();
+      refetch();
+    }
   };
 
   const isSubmittable =
@@ -48,7 +84,7 @@ const AddressForm = ({ address: addressToEdit, user }: AddressFormProps) => {
 
   return (
     <Wrapper>
-      <Form onSubmit={handleSubmit}>
+      <Form>
         <InputSection title="배송지명">
           <ValidationInput
             input={addressName}
@@ -59,7 +95,7 @@ const AddressForm = ({ address: addressToEdit, user }: AddressFormProps) => {
         </InputSection>
         <InputSection title="받는 사람">
           <ValidationInput
-            input={name}
+            input={addressee}
             validation={nameValidation}
             placeholder="이름"
             message={VALIDATION_ERR_MSG.INVALID_NAME}
@@ -74,11 +110,19 @@ const AddressForm = ({ address: addressToEdit, user }: AddressFormProps) => {
           />
         </InputSection>
         <InputSection title="주소">
-          <Address onChangeAddress={handleChangeAddress} />
+          <Address
+            onChangeAddress={handleChangeAddress}
+            defaultAddress={address}
+          />
         </InputSection>
 
         <div className="save-btn">
-          <Button type="submit" size="large" primary disabled={!isSubmittable}>
+          <Button
+            onClick={handleSubmit}
+            size="large"
+            primary
+            disabled={!isSubmittable}
+          >
             저장
           </Button>
         </div>
@@ -107,7 +151,7 @@ const Wrapper = styled.div`
   }
 `;
 
-const Form = styled.form`
+const Form = styled.div`
   max-width: 40rem;
   margin: 3rem auto;
   display: flex;
